Add VariableUtil class
